requirements.txt
cx-Oracle==6.2.1
django-oauth-toolkit==1.1.0
djangorestframework==3.8.2
Pygments==2.2.0
python-decouple==3.1